Move content activity code to fragment
